feat: DATA-2871 Add support for Snowplow globabl contexts
Snowplow's JS tracker accepts global contexts as a way to attach context data to each event. We have
to support deferred loading of context data because of headless implementations, so this change
listens for changes to the ACDL contexts that hold the snowplow global context data and updates
Snowplow when those values change.